requirement.txt
datetime
pickle
threading
bs4
selenium
scrapy.selector 
timeit